scraper/requirements.txt
python-dotenv==1.0.0
azure-servicebus
azure-cosmos
azure-storage-blob===12.19.1
selenium
xmltodict==0.12.0
requests==2.26.0
urllib3==1.26.6
openpyxl==3.0.9
bson==0.5.10